backend/requirements.txt
deepgram-sdk
langchain-groq
langchain-google-genai
fastapi
uvicorn
python-dotenv
langchain
langgraph
google-api-python-client
google-auth
google-auth-oauthlib
pydantic
google-genai
beautifulsoup4
transformers 
torch
pydantic